implementacion crud certificaciones

+++ src/products/services/product.service.ts
@@ -53,13 +53,21 @@ export class ProductService {
       throw new ForbiddenException('No tienes permiso para agregar productos a este negocio.');
     }
 
+    if(business.isActive === false) {
+      throw new ForbiddenException('No puedes agregar productos a un negocio inactivo. Por favor, contacta al administrador para más información.');
+    }
+
+    if (business.status !== 'Active') {
+      throw new ForbiddenException('No puedes agregar productos a un negocio que no está activo. Por favor, espera a que tu negocio sea aprobado por el administrador.');
+    }
+
     const newProduct = this.productRepository.create({
       ...createProductDto,
       business,
     });
 
     const savedProduct = await this.productRepository.save(newProduct);
-    return { message: 'Producto creado exitosamente', product: savedProduct };
+    return { message: `Producto creado exitosamente: ${savedProduct.name}` };
   }
 
   async update(productId: number, updateProductDto: UpdateProductDto, user: any) {
@@ -76,10 +84,18 @@ export class ProductService {
       throw new ForbiddenException('No tienes permiso para editar este producto.');
     }
 
+    if (product.business.isActive === false) {
+      throw new ForbiddenException('No puedes editar productos de un negocio inactivo. Por favor, contacta al administrador para más información.');
+    }
+
+    if (product.business.status !== 'Active') {
+      throw new ForbiddenException('No puedes editar productos de un negocio que no está aprovado. Por favor, contacta al administrador para más información.');
+    }
+
     Object.assign(product, updateProductDto);
     await this.productRepository.save(product);
 
-    return { message: 'Producto actualizado exitosamente', product };
+    return { message: `Producto actualizado exitosamente: ${product.name}`, product };
   }
 
   async remove(productId: number, user: any) {
@@ -94,6 +110,13 @@ export class ProductService {
 
     if (roleName !== 'admin' && product.business.user.id_usuario !== user.id_usuario) {
       throw new ForbiddenException('No tienes permiso para eliminar este producto.');
+    }
+
+    if (product.business.isActive === false) {
+      throw new ForbiddenException('No puedes eliminar productos de un negocio inactivo. Por favor, contacta al administrador para más información.');
+    }
+    if (product.business.status !== 'Active') {
+      throw new ForbiddenException('No puedes eliminar productos de un negocio que no está aprovado. Por favor, contacta al administrador para más información.');
     }
 
     await this.productRepository.remove(product);
